microcluster/types: Remove rest package and merge
This alos consolidates the Server type and replaces it with Status (as it is used for cluster member status)
and instead uses ths Server type to actually specify a server.
