+++ microcluster/types/endpoint.go
@@ -1,4 +1,54 @@
 package types
 
+import (
+	"net/http"
+)
+
+const (
+	// PublicEndpoint - Internally managed APIs.
+	PublicEndpoint EndpointPrefix = "core/1.0"
+
+	// InternalEndpoint - All internal endpoints restricted to trusted servers.
+	InternalEndpoint EndpointPrefix = "core/internal"
+
+	// ControlEndpoint - All internal endpoints available on the local unix socket.
+	ControlEndpoint EndpointPrefix = "core/control"
+)
+
 // EndpointPrefix is a type specifying the endpoint on which the resource exists.
 type EndpointPrefix string
+
+// EndpointAlias represents an alias URL of and Endpoint in our API.
+type EndpointAlias struct {
+	Name string // Name for this alias.
+	Path string // Path pattern for this alias.
+}
+
+// EndpointAction represents an action on an API endpoint.
+type EndpointAction struct {
+	Handler        func(state State, r *http.Request) Response
+	AccessHandler  func(state State, r *http.Request) (trusted bool, resp Response)
+	AllowUntrusted bool
+	ProxyTarget    bool // Allow forwarding of the request to a target if ?target=name is specified.
+}
+
+// Endpoint represents a URL in our API.
+type Endpoint struct {
+	Name    string          // Name for this endpoint.
+	Path    string          // Path pattern for this endpoint.
+	Aliases []EndpointAlias // Any aliases for this endpoint.
+	Get     EndpointAction
+	Put     EndpointAction
+	Post    EndpointAction
+	Delete  EndpointAction
+	Patch   EndpointAction
+
+	AllowedDuringShutdown bool // Whether we should return Unavailable Error (503) if daemon is shutting down.
+	AllowedBeforeInit     bool // Whether we should return Unavailabel Error (503) if the daemon has not been initialized (is not yet part of a cluster).
+}
+
+// Resources represents all the resources served over the same path.
+type Resources struct {
+	PathPrefix EndpointPrefix
+	Endpoints  []Endpoint
+}
